Experimenting with the notion of rendering returning more than a string.
Right now, this is sort of crap.  However, I'm sure there's some merit here.  I htink I need to revisit and work out what the RenderResult object actually is and how it fits together with the rest of the system.  Right now all I've done is make the render process slightly harder to work out and for very little benefit.

Ultimately it should allow an individual renderer to affect the output to the webserver in ways more than just the rendered content. (e.g. headers and result code).